Importing images should work now on 22+

+++ app/src/main/java/me/togaparty/notable_android/data/GalleryImage.kt
@@ -10,7 +10,7 @@ import kotlinx.parcelize.Parcelize
 data class GalleryImage (
         val imageUrl: Uri,
         val name: String,
-        var id: Long = -1,
+        var id: Long = (-1).toLong(),
         var processed: Boolean? = false,
         var textFiles: Map<String,Uri> = linkedMapOf(),
         var wavFiles: Map<String,Uri> = linkedMapOf(),
